+++ CHANGELOG.md
@@ -20,7 +20,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 - language literals were not properly unescaped during Turtle parsing
 - `RDF.Literal.new/1` can take decimals and infers the datatype `xsd:decimal` 
-  correctly  
+  correctly
+- `true` and `false` with capital letters are no longer valid `RDF.Boolean`s 
+  following the XSD specification; the same applies for booleans in Turtle
 
 
 [Compare v0.6.0...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.6.0...HEAD)